Implement new changes for IP-717 by @mjmunger
