+++ application/modules/quotes/models/Mdl_quote_item_amounts.php
@@ -1,12 +1,6 @@
 <?php
 if (!defined('BASEPATH')) exit('No direct script access allowed');
 
-include(__DIR__ . '/QuoteItemInterface.php');
-include(__DIR__ . '/QuoteItemFactory.php');
-include(__DIR__ . '/QuoteItemBase.php');
-include(__DIR__ . '/QuoteItemLegacy.php');
-include(__DIR__ . '/QuoteItem.php');
-
 /*
  * quotePlane
  *
@@ -15,6 +9,8 @@ include(__DIR__ . '/QuoteItem.php');
  * @license		https://quoteplane.com/license.txt
  * @link		https://quoteplane.com
  */
+
+use InvoicePlane\InvoicePlane\ItemFactory;
 
 /**
  * Class Mdl_Quote_Item_Amounts
@@ -29,8 +25,8 @@ class Mdl_Quote_Item_Amounts extends CI_Model
         $this->load->model('quotes/mdl_quote_items');
         $item = $this->mdl_quote_items->get_by_id($item_id);
 
-        $ItemFactory = new QuoteItemFactory();
-        $Item = $ItemFactory->get_item($item);
+        $ItemFactory = new ItemFactory();
+        $Item = $ItemFactory->get_item('quote', $item);
         $db_array = $Item->get_values();
 
         $this->db->where('item_id', $item_id);
