+++ composer.json
@@ -19,5 +19,10 @@
     "codeigniter/framework": "3.1.9",
     "omnipay/omnipay": "~2.0",
     "paragonie/random_compat": "^2.0"
+  },
+  "autoload": {
+    "psr-4": {
+      "InvoicePlane\\InvoicePlane\\": "application/src/"
+    }
   }
 }
